+++ class-01-basic-game-dev/04-0-render-basic/04_triangle_interpolated_render_main.cxx
@@ -45,10 +45,10 @@ int main(int, char**)
     interpolated_render.clear(black);
     interpolated_render.set_gfx_program(program01);
 
-    std::vector<vertex>  triangle_v{ { 0, 0, 1, 0, 0, 0, 0, 0 },
+    std::vector<vertex>   triangle_v{ { 0, 0, 1, 0, 0, 0, 0, 0 },
                                     { 0, 239, 0, 1, 0, 0, 239, 0 },
                                     { 319, 239, 0, 0, 1, 319, 239, 0 } };
-    std::vector<uint8_t> indexes_v{ 0, 1, 2 };
+    std::vector<uint16_t> indexes_v{ 0, 1, 2 };
 
     interpolated_render.draw_triangles(triangle_v, indexes_v);
